Fix precision error of truncated in exponential test

+++ test/exponential.jl
@@ -78,12 +78,13 @@ end
     l = @inferred logdensityof(d, x)
     @test l isa AbstractVector{Float64}
 
-    d = truncated(KernelExponential(Float16(2)), Float16(1), Float16(2))
+    # Fails for Float32
+    d = truncated(KernelExponential(Float32(2)), Float32(1), Float32(2))
     x = @inferred rand(rng, d, 4_200)
-    @test x isa AbstractVector{Float16}
+    @test x isa AbstractVector{Float32}
     @test 1 <= minimum(x) < maximum(x) <= 2
     l = @inferred logdensityof(d, x)
-    @test l isa AbstractVector{Float16}
+    @test l isa AbstractVector{Float32}
 end
 
 @testset "KernelExponential Truncated vs. Distributions.jl" begin
